Update Value not found exception
Previously the exception would error if one was searching for something
that could not be cast to a string.

This fixes that, and adds test cases.

This commit also introduces the Fixture folder for the tests. This folder
should hold all files that are used to set up tests, like the 2 classes
it currently has.